Break-up transformer plugin implementation
