+++ crates/parcel_core/src/types/symbol.rs
@@ -26,7 +26,7 @@ pub struct Symbol {
   /// case in the future
   pub loc: Option<SourceLocation>,
   pub is_weak: bool,
-  /// Only on ESM exports, this may be set to true
+  /// Only on ESM exports, this will be set to true
   pub is_esm_export: bool,
   pub self_referenced: bool,
 }
